cortex/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA

# Embeddings
sentence-transformers>=2.2.0

# API clients
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0

# DNS
dnspython>=2.4.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0

# Web API (Phase 5)
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Tool system
httpx>=0.25.0

# CLI Virtual Machine
rich>=13.0.0
prompt-toolkit>=3.0.40
click>=8.1.0

# Optional: For richer scraping
# beautifulsoup4>=4.12.0
# lxml>=4.6.0
